weather_api/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15.0
redis>=5.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytest
pytest-django